feat: fix compilation on linux
It seems that some of the header files declared duplicate symbols, and
the wrong definitions were being used based on the order of the includes.

By moving one include, I was able to ensure that the right versions
are tried by the compiler.

I also added necessary linker flags to include the EGL libraries on
Linux.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>